+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New Model creation tutorial
 - Added update functionality to the CSV source
 - Added support for Gzip file source
+- Added support for bz2 file source
+- Travis checks for additions to CHANGELOG.md
+- Travis checks for trailing whitespace
 - Added support to pass model specific config options
 ### Changed
 - Restructured documentation to docs folder and moved from rST to markdown